Add new template feature to provisioning
A template is a post-install script that may execute that follows the convention of:

    <system_type>-<group>-<category>-<template>.sh

The <template> part is specified as a new CLI argument: -t <template>.

Example:

    clt provision fedora -g server -c lxc -t db

This would run through a normal provisioning process per the typical rules, except at
the end it will search for: resource/template/fedora-server-lxc-db.sh and execute it.

The provides a mechanism and interface to further provisioning machines to more
specialized needs. (ex. setting up a postgres LXC).