Begin work on testbed command
The testbed command spins up an environment that mocks my homelab.

After all the necessary pieces (example: LXC instances) are created, they are then
provisioned as they are in my lab.

With provisioning completed, a series of tests are performed to confirm that they
are operating correctly.

This testbed feature enables me to update my provisioning process and immediately
run a test to confirm its correctness.