Complete first pass of provision command
Moved all the provision templates and resources out of this project.

Instead of having provisioning templates and resources within this project, they
are expected to be somewhere else in a different folder or project. To give them to
CLT, set the USER_DIR config.

The USER_DIR would be a folder that has a structure like this:

commands/
provision/
  resource/
  templates/
  hosts.yaml
commands.yaml

Note here that commands.yaml can also be there. Essentially, by setting
the USER_DIR, the user can provide not only their provisioning resources and
scripts, but also their own / extra commands.

The contents of USER_DIR/commands.yaml will be merged into CLT's commands.